build: test in diff folder, avoid false positives
I noticed that sometimes the tests could pass, even though generating a
particular contract+frontend combo failed. This could happen when the
frontend had dependencies covered by the root `package.json`, but not
actually included its own `package.json`

Why did this happen? Because we were creating the test project inside
the root folder, so when the test project tried to look up a given
dependency, it was able to find it in the root project's node_modules
folder.

We need to make sure we test in a realistic environment, where the
project cannot depend on the root folder's node_modules