fix: assemblyscript install with npm@7 (node@15)
fixes #533

npm@7 ships with node@15 and changes behavior to automatically install
peerDependencies. It seems that the peerDependencies for near-sdk-as
aren't configured perfectly, and this causes problems. Once we figure
out this upstream problem, we may also need to add more explicit
dependencies at the root of the project, since npm@7 may not install
dependencies-of-dependencies to the root node_modules folder, either.

This will be addressed in https://github.com/near/near-sdk-as/issues/312

https://blog.npmjs.org/post/626173315965468672/npm-v7-series-beta-release-and-semver-major